8272745: VectorAPI: Vector/Mask type cleanup for masked vector intrinsics
Some of the VectorAPIs were inconsistent with their intrinsic
declarations, and some of the type arguments were not precise enough as
well. E.g., The return type of "unaryOp" should be a subtype of "Vector"
instead of an all-purpose "Object" [1][2].

The imprecise declarations and type arguments would not effect the
correctness but bring bad readability and make the code hard to
maintain. This patch is a cleanup for those mismatched declarations of
VectorAPIs.

[1] https://github.com/openjdk/panama-vector/blob/vectorIntrinsics%2Bmask/src/java.base/share/classes/jdk/internal/vm/vector/VectorSupport.java#L269
[2] https://github.com/openjdk/panama-vector/blob/vectorIntrinsics%2Bmask/src/hotspot/share/classfile/vmIntrinsics.hpp#L812

Change-Id: I445cef5f7d778e431479b86b99f6e3b968ed1148