Fix a predicate register allocation issue
When a predicate live range has a predicate register mask chosen, we
create a map between node and opto reg. For scalable register, when it
is allocated at stack, we need to record the real hi/lo opto reg  pair
to indicate the size of stack it occupied. But for a physical register
allocated, we just mask the whole reg mask slots, e.g. only one slot
for predicate register. This patch fixes the issue for predicate
register handling. When predicate is allocated as a physical register,
e.g. 335 (P15), we should not call set2(), which will assign both 335
and 336 for this live range.

Also move idealreg2spillmask[Op_RegVectMask] initialization code into
scalable vector handling block to align with the handling for Op_VecA.

Change-Id: I64b42f654a158c2f40a0b6792740678d07aac933