Optimize long range checks Try to optimize long RCE by adding loop predicates, for loops using Long RC (`CmpUL`).
This code is intended to handle int counted loops with Long RC, so
as well striped loops.
* modify `is_scaled_iv_plus_offset` to allow searching for induction
  variables over _long_ nodes
* modify `rc_predicate` to add overflow checks if long could overflow
* and bail-out conditions to  `PhaseIdealLoop::split_thru_phi` to prevent
   this optimizations (this code would not be optimized otherwise
   `for (long l=2; l < length; l += 4) Objects.checkIndex(2 + l, length);`
* prevent long RCE when scale is larger than `1 << 16`,
   for now there's not overflow check code;
This is inital commit, just to validate if this is a good approach, and
checks are ok.

Especially worth for double checks are some assumptions about overflow.

Later we could:
- better handle SubL, scale, and multiplications
in `is_scaled_iv_plus_offset`
- create extended Add, Mul, Sub projecting value and overflow
(or better all CR flags) - will increase performance slightly,
- better optimize Overflow nodes `Ideal` in IGVN