8285013: AArch64: [vectorapi] Backend support of ExpandV for SVE2
This patch adds the backend for Vector.expand(mask).

The basic idea is to use TBL which can shuffle the elements in the given
vector flexibly. HISTCNT is used to generate the second source input for
TBL whose value is used to select the indexed element from src vector.

The final code for an int vector is generated as below:

```
   mov z17.s, #0
   histcnt z17.s, p0/z, z17.s, z17.s
   sub z17.s, z17.s, #1
   tbl z17.s, {z16.s}, z17.s
```

As the HISTCNT instruction only supports element in size of 32/64 bits,
expanding of subword type is missing now.

Change-Id: I0f08efcf4a740ee6719ab77bfba9a923e9c41eed